+++ packages/spacecat-shared-data-access/test/unit/v2/util/patcher.test.js
@@ -85,6 +85,7 @@ describe('Patcher', () => {
         indexes: mockEntity.model.indexes,
         model: mockEntity.model,
         references: [],
+        options: { allowRemove: true, allowUpdates: true },
       },
     );
 
@@ -99,6 +100,12 @@ describe('Patcher', () => {
     patcher.patchValue('name', 'UpdatedName');
     expect(mockEntity.patch().set.calledWith({ name: 'UpdatedName' })).to.be.true;
     expect(mockRecord.name).to.equal('UpdatedName');
+  });
+
+  it('throws error if schema prhibits updates', () => {
+    patcher.schema.options.allowUpdates = false;
+    expect(() => patcher.patchValue('name', 'UpdatedName'))
+      .to.throw('Updates prohibited by schema for MockEntityModel.');
   });
 
   it('throws error for read-only property', () => {
@@ -142,6 +149,12 @@ describe('Patcher', () => {
 
     expect(mockEntity.patch().go.calledOnce).to.be.true;
     expect(isIsoDate(mockRecord.updatedAt)).to.be.true;
+  });
+
+  it('throws error when saving with updates prohibited by schema', async () => {
+    patcher.schema.options.allowUpdates = false;
+
+    expect(patcher.save()).to.be.rejectedWith('Updates prohibited by schema for MockModel.');
   });
 
   it('does not save if there are no updates', async () => {